+++ src/OpeningHours.php
@@ -204,4 +204,63 @@ class OpeningHours
 
         return $date;
     }
+
+    public function map(callable $callback): array
+    {
+        return Arr::map($this->openingHours, $callback);
+    }
+
+    public function flatMap(callable $callback): array
+    {
+        return Arr::flatMap($this->openingHours, $callback);
+    }
+
+    public function mapExceptions(callable $callback): array
+    {
+        return Arr::map($this->exceptions, $callback);
+    }
+
+    public function flatMapExceptions(callable $callback): array
+    {
+        return Arr::flatMap($this->exceptions, $callback);
+    }
+
+    public function asStructuredData(): array
+    {
+        $regularHours = $this->flatMap(function (OpeningHoursForDay $openingHoursForDay, string $day) {
+            return $openingHoursForDay->map(function (TimeRange $timeRange) use ($day) {
+                return [
+                    '@type' => 'OpeningHoursSpecification',
+                    'dayOfWeek' => ucfirst($day),
+                    'opens' => (string) $timeRange->start(),
+                    'closes' => (string) $timeRange->end(),
+                ];
+            });
+        });
+
+        $exceptions = $this->flatMapExceptions(function (OpeningHoursForDay $openingHoursForDay, string $date) {
+
+            if ($openingHoursForDay->isEmpty()) {
+                return [[
+                    '@type' => 'OpeningHoursSpecification',
+                    'opens' => '00:00',
+                    'closes' => '00:00',
+                    'validFrom' => $date,
+                    'validThrough' => $date,
+                ]];
+            }
+
+            return $openingHoursForDay->map(function (TimeRange $timeRange) use ($date) {
+                return [
+                    '@type' => 'OpeningHoursSpecification',
+                    'opens' => $timeRange->start(),
+                    'closes' => $timeRange->end(),
+                    'validFrom' => $date,
+                    'validThrough' => $date,
+                ];
+            });
+        });
+
+        return array_merge($regularHours, $exceptions);
+    }
 }