+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All Notable changes to `opening-hours` will be documented in this file
 
+## 1.2.0 - 2017-01-03
+- Added `asStructuredData` to retrieve the opening hours as a Schema.org structured data array 
+- Added utility methods: `OpeningHours::map`, `OpeningHours::flatMap`, `OpeningHours::mapExceptions`, `OpeningHours::flatMapExceptions`,`OpeningHoursForDay::map` and `OpeningHoursForDay::empty`
+
 ## 1.1.0 - 2016-11-09
 - Added timezone support